feat: Periodically clear away dead ShardReaders (#13)
* Check for and remove dead threads on every loop of Kinesis::Consumer@lock_duration
* Add report_on_exception to the reader thread so something is logged if it dies unexpectedly

report_on_exception simply logs to stderr if the thread dies. An alternative might be abort_on_exception which will reraise the exception in the main thread. I decided to go with the less disruptive option, because there is the alternative case where no exception is raised but the thread stops naturally, and the solution of clearing dead ShardReaders in the loop in Consumer covers both cases.